+++ src/main/java/org/tugraz/sysds/runtime/io/WriterMatrixMarketParallel.java
@@ -72,11 +72,11 @@ public class WriterMatrixMarketParallel extends WriterMatrixMarket
 			int blklen = (int)Math.ceil((double)rlen / numThreads);
 			for(int i=0; i<numThreads & i*blklen<rlen; i++) {
 				Path newPath = new Path(path, IOUtilFunctions.getPartFileName(i));
-				tasks.add(new WriteMMTask(newPath, job, fs, src, i*blklen, (int)Math.min((i+1)*blklen, rlen)));
+				tasks.add(new WriteMMTask(newPath, job, fs, src, i*blklen, Math.min((i+1)*blklen, rlen)));
 			}
 
 			//wait until all tasks have been executed
-			List<Future<Object>> rt = pool.invokeAll(tasks);	
+			List<Future<Object>> rt = pool.invokeAll(tasks);
 			pool.shutdown();
 			
 			//check for exceptions
